[css-text] There is no shape-result when items overflow under break-word
When the overflow-wrap: 'break-word' feature is used, items overflowing
don't have an associated ShapeResult instance, since we deliberately set
the kNoResultIfOverflow to ignore it because it will be rewound in any
case.

However, in the RewindOverflow function, we incorrectly assumed that
when an item has no associated ShapeResult, it shouldn't be a trailable
text item.

Additionally, this CL removes the clause that implied that items formed
with all trailable spaces should be ignored. Such logic prevented such
kind of items to hang, which is what the spec mandates when pre-wrap is
being used.

Bug: 1003184
Change-Id: Iab930be3ecc8e4dc814be25d902ecedabc70a944