Set timeout=long for FileSystemBaseHandle-postMessage.js tests
This is like the other NativeFileSystem postMessage() tests recently
updated to use a long timeout.  These tests are similiar to the
previously updated tests except that they use message ports instead of
direct messaging.  The test history dashboard shows that these tests
take 5 or more seconds to complete, which is close to the 7 second
timeout.

This change adds "meta timeout=long" to both test variants to extend
the 7 second timeout to a longer value.

Bug: 1044519
Change-Id: I87d14e819f7d7813a3708695a3899271d2bdf620